requirements.txt
trafilatura
qdrant-client
httpx
python-dotenv
pdfplumber
tiktoken
rich
lxml_html_clean
